pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Compare two JSON or YML files"
authors = ["NRomanLee <leeroman680@gmail.com>"]
readme = "README.md"
packages = [{ include = "gendiff" }]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
coverage = "^5.2.1" 
pytest-cov = "^5.0.0"
requests = "^2.25.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
flake8 = "^7.1.0"
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"